changed id to class
I changed the id: "fName" to a class so I can place the name throughout the website. Also changed the function to create a variable of the class then run that variable through a loop to present the value in every instance. same thing done to the dipslayText function for the "hidden" class